refactor(all)!: merapikan hasil brainstrom
